+++ src/compiler/bundle/rollup-plugins/local-resolution.ts
@@ -1,10 +1,10 @@
-import { Config } from '../../../declarations';
+import { CompilerCtx, Config } from '../../../declarations';
 import { normalizePath } from '../../util';
 
 
-export default function localResolver(config: Config) {
+export default function localResolution(config: Config, compilerCtx: CompilerCtx) {
   return {
-    name: 'localResolverPlugin',
+    name: 'localResolution',
 
     async resolveId(importee: string, importer: string) {
       importee = normalizePath(importee);
@@ -28,14 +28,13 @@ export default function localResolver(config: Config) {
       const dirIndexFile = config.sys.path.join(directory + importee, 'index.js');
 
       let stats;
-
       try {
-        stats = config.sys.fs.statSync(importee);
+        stats = await compilerCtx.fs.stat(dirIndexFile);
       } catch (e) {
         return null;
       }
 
-      if (stats.isFile()) {
+      if (stats.isFile) {
         return dirIndexFile;
       }