+++ src/compiler/config/validate-config.ts
@@ -4,6 +4,8 @@ import { validateCopy } from './validate-copy';
 import { validateDependentCollection } from './validate-collection';
 import { validateNamespace } from './validate-namespace';
 import { validatePaths } from './validate-paths';
+import { validatePlugins } from './validate-plugins';
+import { validatePublicPath } from './validate-public-path';
 
 
 export function validateBuildConfig(config: Config, setEnvVariables?: boolean) {
@@ -42,6 +44,9 @@ export function validateBuildConfig(config: Config, setEnvVariables?: boolean) {
   // figure out all of the config paths and absolute paths
   validatePaths(config);
 
+  // figure out the client-side public path
+  validatePublicPath(config);
+
   // default devMode false
   config.devMode = !!config.devMode;
 
@@ -73,6 +78,8 @@ export function validateBuildConfig(config: Config, setEnvVariables?: boolean) {
   setBooleanConfig(config, 'generateWWW', true);
 
   validateCopy(config);
+
+  validatePlugins(config);
 
   if (!config.watchIgnoredRegex) {
     config.watchIgnoredRegex = DEFAULT_WATCH_IGNORED_REGEX;