+++ src/compiler/transpile/transformers/component-dependencies.ts
@@ -1,4 +1,4 @@
-import { BuildCtx, CompilerCtx } from '../../../declarations';
+import { BuildCtx, CompilerCtx, PotentialComponentRef } from '../../../declarations';
 import { MEMBER_TYPE } from '../../../util/constants';
 import { normalizePath } from '../../util';
 import * as ts from 'typescript';
@@ -6,19 +6,14 @@ import * as ts from 'typescript';
 
 export function componentDependencies(compilerCtx: CompilerCtx, buildCtx: BuildCtx): ts.TransformerFactory<ts.SourceFile> {
 
-  const allComponentTags = Object.keys(compilerCtx.moduleFiles)
-    .map(filePath => compilerCtx.moduleFiles[filePath].cmpMeta)
-    .filter(cmpMeta => cmpMeta && cmpMeta.tagNameMeta)
-    .map(cmpMeta => cmpMeta.tagNameMeta);
-
   return (transformContext) => {
 
     function visit(node: ts.Node, filePath: string): ts.VisitResult<ts.Node> {
       if (node.kind === ts.SyntaxKind.CallExpression) {
-        callExpression(buildCtx, allComponentTags, filePath, node as ts.CallExpression);
+        callExpression(buildCtx, filePath, node as ts.CallExpression);
 
       } else if (node.kind === ts.SyntaxKind.StringLiteral) {
-        stringLiteral(buildCtx, allComponentTags, filePath, node as ts.StringLiteral);
+        stringLiteral(buildCtx, filePath, node as ts.StringLiteral);
       }
 
       return ts.visitEachChild(node, (node) => {
@@ -27,15 +22,17 @@ export function componentDependencies(compilerCtx: CompilerCtx, buildCtx: BuildC
     }
 
     return (tsSourceFile) => {
-      addPropConnects(compilerCtx, buildCtx, tsSourceFile.fileName);
+      const filePath = normalizePath(tsSourceFile.fileName);
 
-      return visit(tsSourceFile, tsSourceFile.fileName) as ts.SourceFile;
+      addPropConnects(compilerCtx, buildCtx.componentRefs, filePath);
+
+      return visit(tsSourceFile, filePath) as ts.SourceFile;
     };
   };
 }
 
 
-function addPropConnects(compilerCtx: CompilerCtx, buildCtx: BuildCtx, filePath: string) {
+function addPropConnects(compilerCtx: CompilerCtx, sourceStrings: PotentialComponentRef[], filePath: string) {
   const moduleFile = compilerCtx.moduleFiles[filePath];
 
   const cmpMeta = (moduleFile && moduleFile.cmpMeta);
@@ -47,7 +44,7 @@ function addPropConnects(compilerCtx: CompilerCtx, buildCtx: BuildCtx, filePath:
     Object.keys(cmpMeta.membersMeta).forEach(memberName => {
       const memberMeta = cmpMeta.membersMeta[memberName];
       if (memberMeta.memberType === MEMBER_TYPE.PropConnect) {
-        buildCtx.componentRefs.push({
+        sourceStrings.push({
           tag: memberMeta.ctrlId,
           filePath: filePath
         });
@@ -57,65 +54,47 @@ function addPropConnects(compilerCtx: CompilerCtx, buildCtx: BuildCtx, filePath:
 }
 
 
-function callExpression(buildCtx: BuildCtx, allComponentTags: string[], filePath: string, node: ts.CallExpression) {
-  if (node && node.arguments) {
+function callExpression(buildCtx: BuildCtx, filePath: string, node: ts.CallExpression) {
+  if (node.arguments && node.arguments[0]) {
 
     if (node.expression.kind === ts.SyntaxKind.Identifier) {
       // h('tag')
-      callExpressionArg(buildCtx, allComponentTags, filePath, node.expression as ts.Identifier, node.arguments);
+      callExpressionArg(buildCtx, filePath, node.expression as ts.Identifier, node.arguments);
 
     } else if (node.expression.kind === ts.SyntaxKind.PropertyAccessExpression) {
       // document.createElement('tag')
-
       if ((node.expression as ts.PropertyAccessExpression).name) {
         // const
-        callExpressionArg(buildCtx, allComponentTags, filePath, (node.expression as ts.PropertyAccessExpression).name as ts.Identifier, node.arguments);
+        callExpressionArg(buildCtx, filePath, (node.expression as ts.PropertyAccessExpression).name as ts.Identifier, node.arguments);
       }
     }
   }
 }
 
 
-function callExpressionArg(buildCtx: BuildCtx, allComponentTags: string[], filePath: string, callExpressionName: ts.Identifier, args: ts.NodeArray<ts.Expression>) {
+function callExpressionArg(buildCtx: BuildCtx, filePath: string, callExpressionName: ts.Identifier, args: ts.NodeArray<ts.Expression>) {
   if (TAG_CALL_EXPRESSIONS.includes(callExpressionName.escapedText as string)) {
 
-    if (args[0] && args[0].kind === ts.SyntaxKind.StringLiteral) {
-      addComponentReference(buildCtx, allComponentTags, filePath, (args[0] as ts.StringLiteral).text);
+    if (args[0].kind === ts.SyntaxKind.StringLiteral) {
+      const tag = (args[0] as ts.StringLiteral).text;
+
+      if (typeof tag === 'string') {
+        buildCtx.componentRefs.push({
+          tag: tag,
+          filePath: filePath
+        });
+      }
     }
   }
 }
 
 
-function stringLiteral(buildCtx: BuildCtx, allComponentTags: string[], filePath: string, node: ts.StringLiteral) {
-  let t = node.text;
-
-  if (typeof t === 'string' && t.includes('<')) {
-    t = t.toLowerCase()
-         .replace(/\s/g, '~');
-
-    const foundTags = allComponentTags
-      .filter(tag => {
-        return t.includes('<' + tag + '>') ||
-               t.includes('<' + tag + '~');
-      });
-
-    foundTags.forEach(foundTag => {
-      addComponentReference(buildCtx, allComponentTags, filePath, foundTag);
+function stringLiteral(buildCtx: BuildCtx, filePath: string, node: ts.StringLiteral) {
+  if (typeof node.text === 'string' && node.text.includes('<')) {
+    buildCtx.componentRefs.push({
+      html: node.text,
+      filePath: filePath
     });
-  }
-}
-
-
-function addComponentReference(buildCtx: BuildCtx, allComponentTags: string[], filePath: string, tag: string) {
-  if (typeof tag === 'string') {
-    tag = tag.toLowerCase().trim();
-
-    if (allComponentTags.includes(tag)) {
-      buildCtx.componentRefs.push({
-        tag: tag,
-        filePath: normalizePath(filePath)
-      });
-    }
   }
 }